requirements.txt
fastapi
uvicorn
httpx
pydantic
google-generativeai
python-dotenv
supabase
pinecone
qdrant-client
fastembed
gunicorn
